requirements.txt
pandas>=2.2
numpy>=1.26
pyarrow>=16.0
yfinance>=0.2.40
pandas-market-calendars>=4.4.1
pandas-datareader>=0.10.0
pyyaml>=6.0
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
cvxpy>=1.4.0

